+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/toolWindow/components/ProcessTablePanel.kt
@@ -12,27 +12,35 @@ class ProcessTablePanel : BorderLayoutPanel() {
     private lateinit var processName: Cell<JLabel>
     private lateinit var filename: Cell<JLabel>
     private lateinit var startTime: Cell<JLabel>
+    private lateinit var commandLine: Cell<JLabel>
+    private lateinit var operatingSystem: Cell<JLabel>
+    private lateinit var processArchitecture: Cell<JLabel>
 
     init {
         border = JBUI.Borders.empty(5, 10)
 
         val panel = panel {
             group("Process Details") {
-                row {
-                    label("Id:").bold()
+                row("Id:") {
                     processId = label("-")
                 }
-                row {
-                    label("Name:").bold()
+                row("Name:") {
                     processName = label("-")
                 }
-                row {
-                    label("Filename:").bold()
+                row("Filename:") {
                     filename = label("-")
                 }
-                row {
-                    label("Start time:").bold()
+                row("Start time:") {
                     startTime = label("-")
+                }
+                row("Command line:") {
+                    commandLine = label("-")
+                }
+                row("Operating system:") {
+                    operatingSystem = label("-")
+                }
+                row("Process architecture:") {
+                    processArchitecture = label("-")
                 }
             }
         }
@@ -45,6 +53,9 @@ class ProcessTablePanel : BorderLayoutPanel() {
         processName.component.text = processInfo.processName
         filename.component.text = processInfo.filename
         startTime.component.text = processInfo.startTime
+        commandLine.component.text = processInfo.commandLine
+        operatingSystem.component.text = processInfo.operatingSystem
+        processArchitecture.component.text = processInfo.processArchitecture
     }
 
     fun clear() {
@@ -52,5 +63,8 @@ class ProcessTablePanel : BorderLayoutPanel() {
         processName.component.text = "-"
         filename.component.text = "-"
         startTime.component.text = "-"
+        commandLine.component.text = "-"
+        operatingSystem.component.text = "-"
+        processArchitecture.component.text = "-"
     }
 }